Revert "New updating algorithms added"
