+++ Documentation/Settings_in_Optimization_Dict.md
@@ -290,11 +290,12 @@ Please see [this example](https://github.com/Quantum-OCS/QuOCS/blob/develop/Docu
 
 | Setting | Type | Explanation |
 | --- | --- | --- |
-|**"basis_name"** |*string*| Name of the basis in which to expand the pulse. So far on can select from "Fourier", "Chebyshev", "PiecewiseBasis", and "Sigmoid". Depending on the basis, the randomized super-parameter differs. It is recommended to start with the Fourier basis. In that case the randomly selected parameter is the frequency of the trigonometrical function updating the pulse during dCRAB iterations. For more information please contact the developers. |
+|**"basis_name"** |*string*| Name of the basis in which to expand the pulse. So far on can select from "Fourier", "Chebyshev", "PiecewiseBasis", "Sinc", and "Sigmoid". Depending on the basis, the randomized super-parameter differs. It is recommended to start with the Fourier basis. In that case the randomly selected parameter is the frequency of the trigonometrical function updating the pulse during dCRAB iterations. For more information please contact the developers. |
 |**"basis_vector_number"** |*Int*| Number of vectors used for basis expansion i each super-iteration of the dCRAB algorithm. Typically, one vector corresponds to 1-2 parameters to optimize (e.g. amplitude and phase of a sine function). The higher this value is set, the longer each SI takes to converge. For more information please contact the developers. *(Default: 1)* |
 |**"random_super_parameter_distribution"** |*dict*| Distribution from which to sample the randomized super-parameter. So far, the only option is "distribution_name": "Uniform" where one can set to upper and lower limits of the parameter. In the case of the Fourier basis, this corresponds to the number of oscillations allowed during the pulse time. Therefore, depending on the length of the pulse, this enforces bandwidth constraints. For more information please contact the developers. |
 |**"sigma"** *(optional)* |*float*| The Sigmoid basis is designed to keep the pulse spectrum envelope-limited. Sigma specifies the steepness of the individual rises, which will in turn reduce high-frequency excitations. A good point to start is final_time/100. *(Default: 0.1)* |
 |**"offset"** *(optional)* |*float*| The Sigmoid basis is designed to keep the pulse spectrum envelope-limited and the pulse zero at beginning and end. The offset determines how far from the edges it will go down. This parameter should depend on sigma and the maximum amplitude. A good point to start is sigma x (upper_limit - loweer_limit)/10. *(Default: 0.1)* |
+| **"basis_max"** *(optional)*              |*float*| The Sinc basis keeps a spectrum of consistent width defined by this parameter. The units are defined in the same way as the Fourier frequencies.  *(Default: 1.)*|
 
 ## Bases Overview
 | Basis name | Implementation | Visualization |
