+++ Examples/settings_dCRAB.json
@@ -4,11 +4,11 @@
     "algorithm_settings": {
         "algorithm_name": "dCRAB",
         "super_iteration_number": 3,
-        "max_eval_total": 500,
+        "max_eval_total": 3000,
         "optimization_direction": "minimization",
         "dsm_settings": {
             "general_settings": {
-                "dsm_algorithm_name": "CMAES"
+                "dsm_algorithm_name": "NelderMead"
             },
             "stopping_criteria": {
                 "xatol": 1e-4,
