+++ Other/docs/html/_modules/index.html
@@ -114,7 +114,6 @@
 <li><a href="quocslib/communication/AllInOneCommunication.html">quocslib.communication.AllInOneCommunication</a></li>
 <li><a href="quocslib/figureofmeritevaluation/AbstractFoM.html">quocslib.figureofmeritevaluation.AbstractFoM</a></li>
 <li><a href="quocslib/gradientfreemethods/CMAES.html">quocslib.gradientfreemethods.CMAES</a></li>
-<li><a href="quocslib/gradientfreemethods/CMAES2.html">quocslib.gradientfreemethods.CMAES2</a></li>
 <li><a href="quocslib/gradientfreemethods/DirectSearchMethod.html">quocslib.gradientfreemethods.DirectSearchMethod</a></li>
 <li><a href="quocslib/gradientfreemethods/GradientFreeTemplate.html">quocslib.gradientfreemethods.GradientFreeTemplate</a></li>
 <li><a href="quocslib/gradientfreemethods/NelderMead.html">quocslib.gradientfreemethods.NelderMead</a></li>
